Settings up Modals
Changed the old method for Medals inside the event command. Didn't changed for the two others yet.

Fixed a error when the remove or edit command was trying to get the Event ID variable.

Removed useless package sqlite.

Added a comment at the line 67 inside commands/event.js about a issues with the database system.