Update player.js
Created a list system for the player display sub command.

Issues inside player.js inside display subcommand. The index and number of user displayed per page change from one to two bigger or smaller. I was unable to find the cause of the error. I would consider the error as minor but we still need to fix it.